+++ hyrisecockpit/frontend/src/components/alerts/LinearLoader.vue
